+++ VTOS.Application/Features/Users/DTOs/UpdateAvatarRequest.cs
@@ -1,0 +1,5 @@
+﻿using Microsoft.AspNetCore.Http;
+namespace VTOS.Application.Features.Users.Commands
+{
+    public record UpdateAvatarRequest(IFormFile Avatar);
+}