+++ VTOS.Application/Features/Users/DTOs/UpdateProfileRequest.cs
@@ -1,0 +1,11 @@
+﻿using VTOS.Domain.Enums;
+
+namespace VTOS.Application.Features.Users.DTOs
+{
+    public record UpdateProfileRequest
+    (
+        string? FullName,
+        DateTime? DOB,
+        Gender? Gender
+    );
+}